fix subcategory and fix bug

+++ database/migrations/2023_05_04_071752_tbl_order.php
@@ -14,7 +14,7 @@ class TblOrder extends Migration
     public function up()
     {
         Schema::create('tbl_order', function (Blueprint $table) {
-            $table->bigIncrements('order_id');
+            $table->integer('order_id');
             $table->integer('customer_id');
             $table->integer('shipping_id');
             $table->string('order_total');
